+++ src/js/services/toast/toast.module.js
@@ -1,0 +1,5 @@
+import service from "./toast.service";
+
+angular
+.module("uploader.toast", [])
+.factory("Toast", service);