+++ CHANGELOG.md
@@ -1,16 +1,18 @@
 # Changelog
 
-### 0.2.2
+### 0.3.0
 
 **Minor**
 
-- none
+- Added `electron-debug`
 
 **Patch**
 
 - split angular files a little more
 - renamed Ajax and MainCtrl to API and MainController
 - general John Papa styleguide stuff
+- wrap toast calls in a service
+- toast templates in separate files
 
 ### 0.2.1